+++ src/qcoder/__init__.py
@@ -1,3 +1,3 @@
 __all__ = []
-__version__ = "0.5.0a6"
+__version__ = "0.5.0a8"
 file = __file__